+++ kite-spring-boot3-autoconfigure/src/main/kotlin/com/tang/kite/boot/autoconfigure/SpringBoot3KiteProperties.kt
@@ -32,3 +32,7 @@ class SpringBoot3TenantProperties : TenantProperties()
 @Suppress("ConfigurationProperties")
 @ConfigurationProperties(prefix = DataSourceProperties.DATA_SOURCE_PREFIX)
 class SpringBoot3DataSourceProperties : DataSourceProperties()
+
+@Suppress("ConfigurationProperties")
+@ConfigurationProperties(prefix = SchemaProperties.SCHEMA_PREFIX)
+class SpringBoot3SchemaProperties : SchemaProperties()